+++ Zepholume-Shaders/shaders/lib/cloud_vertex.glsl
@@ -1,6 +1,7 @@
 #ifndef ZEPHO_CLOUD_VERTEX_GLSL
 #define ZEPHO_CLOUD_VERTEX_GLSL
 #include "/lib/profile.glsl"
+#include "/lib/common.glsl"
 varying vec2 zephCloudTexCoord;
 varying vec4 zephCloudVertexColour;
 varying float zephCloudDistance;
@@ -11,7 +12,7 @@ void main() {
     zephCloudTexCoord = gl_MultiTexCoord0.st;
     zephCloudVertexColour = gl_Color;
     zephCloudDistance = length(viewPosition.xyz);
-    zephCloudNormalUp = clamp((gl_NormalMatrix * gl_Normal).y * 0.5 + 0.5, 0.0, 1.0);
+    zephCloudNormalUp = clamp(zephSafeNormalize(gl_NormalMatrix * gl_Normal).y * 0.5 + 0.5, 0.0, 1.0);
     gl_Position = gl_ProjectionMatrix * viewPosition;
 }
 #endif